Split the ledger builders out of continuityGuardrail.ts
The guardrail file had grown to 514 lines once the assertion, commitment,
and scene-change builders landed. Move the ledger half into a pure
continuityLedger.ts (still no store imports) and leave detection,
formatting, and the correction prompt where they were. No behavior change;
the 47 continuity and extractor tests pin it.
